연습용 게시판 CRUD완료

+++ app/src/main/java/com/goteatfproject/appgot/web/MyController.java
@@ -1,12 +1,19 @@
 package com.goteatfproject.appgot.web;
 
+import com.goteatfproject.appgot.vo.AttachedFile;
+import com.goteatfproject.appgot.vo.Party;
 import java.io.File;
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import javax.servlet.ServletContext;
+import javax.servlet.ServletException;
 import javax.servlet.http.HttpSession;
+import javax.servlet.http.Part;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -40,16 +47,12 @@ public class MyController {
   FeedService feedService;
   @Autowired
   MemberService memberService;
-
   @Autowired
   EventService eventService;
-
   @Autowired
   FollowerService followerService;
-
   @Autowired
   BoardService boardService;
-
   @Autowired
   ServletContext sc;
 
@@ -98,7 +101,7 @@ public class MyController {
     System.out.println("member = " + member);
     System.out.println(member.getPassword() == "");
     // 새로운 패스워드가 없을때는 udpate2()
-    if(member.getPassword() == "") {
+    if (member.getPassword() == "") {
       memberService.update2(member);
     } else { // 새로운 패스워드 변경이 있을때 update()
       memberService.update(member);
@@ -109,7 +112,8 @@ public class MyController {
 
   // 마이페이지 메인 프로필 사진 수정
   @PostMapping("/updateProfile")
-  public String updateProfile(@RequestParam("files") MultipartFile files, HttpSession session) throws Exception {
+  public String updateProfile(@RequestParam("files") MultipartFile files, HttpSession session)
+      throws Exception {
     Member member = (Member) session.getAttribute("loginMember");
     if (!files.isEmpty()) {
       String dirPath = sc.getRealPath("/member/files");
@@ -140,14 +144,21 @@ public class MyController {
   // 마이페이지 회원 비활성화
   @PostMapping("/delete")
   @ResponseBody
-  public String delete(HttpSession session, int no) throws Exception {
-    Member member = (Member) session.getAttribute("loginMember");
-    if (member.getNo() == no) {
-      memberService.delete(no);
-      session.invalidate();
-      return "회원 탈퇴 완료";
+  public String delete(HttpSession session, @RequestParam("no") int no,
+      @RequestParam("outState") boolean outState) throws Exception {
+
+    System.out.println("out =====> " + outState);
+
+    if (outState != false) {
+      Member member = (Member) session.getAttribute("loginMember");
+      if (member.getNo() == no) {
+        memberService.delete(no);
+        session.invalidate();
+        return "1";
+//        return "회원 탈퇴 완료";
+      }
     }
-    return "회원 탈퇴 실패";
+    return "0";
   }
 
   // 마이페이지-파티게시글 관리
@@ -189,7 +200,7 @@ public class MyController {
   public String partyDeletes(@RequestParam("checkedValue[]") int[] checkedValue) throws Exception {
     int valueLength = checkedValue.length;
 
-    for(int i=0; i < valueLength; i++) {
+    for (int i = 0; i < valueLength; i++) {
       System.out.println(checkedValue[i]);
       partyService.allDelete(checkedValue[i]);
     }
@@ -254,7 +265,7 @@ public class MyController {
   public String feedDeletes(@RequestParam("checkedValue[]") int[] checkedValue) throws Exception {
     int valueLength = checkedValue.length;
 
-    for(int i=0; i < valueLength; i++) {
+    for (int i = 0; i < valueLength; i++) {
       System.out.println(checkedValue[i]);
       feedService.allDelete2(checkedValue[i]);
     }
@@ -271,7 +282,6 @@ public class MyController {
     System.out.println("feedAttachedFile.getNo() = " + feedAttachedFile.getNo());
     System.out.println("feedAttachedFile.getNo() = " + feedAttachedFile.getFilepath());
     System.out.println("feedAttachedFile.getNo() = " + feedAttachedFile.getFeedNo());
-
 
     // 게시글 작성자 일치여부
     Member loginMember = (Member) session.getAttribute("loginMember");
@@ -352,7 +362,7 @@ public class MyController {
 
     Member loginMember = (Member) session.getAttribute("loginMember");
 
-    if(loginMember != null) {
+    if (loginMember != null) {
       List<Follower> followList = followerService.selectFollowList(loginMember.getNo());
       model.addAttribute("follows", followList);
       System.out.println("model.getAttribute(\"follows\") = " + model.getAttribute("follows"));
@@ -373,11 +383,86 @@ public class MyController {
   public String followDeletes(@RequestParam("checkedValue[]") int[] checkedValue) throws Exception {
     int valueLength = checkedValue.length;
 
-    for(int i=0; i < valueLength; i++) {
+    for (int i = 0; i < valueLength; i++) {
       System.out.println(checkedValue[i]);
       followerService.allDelete3(checkedValue[i]);
     }
     return "삭제 성공";
+  }
+
+  // 파티 게시물 수정
+  @PostMapping("updateParty")
+  public String update(Party party, HttpSession session,
+      Part[] files) throws Exception {
+
+    party.setAttachedFiles(saveAttachedFiles(files));
+
+    // detail.html : <input name="no" type="number" value="1" th:value="${party.no}" readonly hidden/>
+    // 위에 추가해야 party.getNo() 가져오기 가능 System.out.println("partyNo = " + party.getNo());
+    checkOwner2(party.getNo(), session);
+
+    if (!partyService.update(party)) {
+      throw new Exception("게시글을 변경할 수 없습니다.");
+    }
+    return "redirect:myPartyList";
+  }
+
+  private void checkOwner2(int partyNo, HttpSession session) throws Exception {
+    Member loginMember = (Member) session.getAttribute("loginMember");
+    // 개인이해메모
+    // getWriter().getNo() != loginMember.getNo() // 로그인 멤버no 꺼내서 party에 있는 Member writer 이용해서 일치여부 확인
+    // 방향 ----->
+    if (partyService.get(partyNo).getWriter().getNo() != loginMember.getNo()) {
+      throw new Exception("파티 게시글 작성자가 아닙니다.");
+    }
+  }
+
+  private List<AttachedFile> saveAttachedFiles(Part[] files)
+      throws IOException, ServletException {
+    List<AttachedFile> attachedFiles = new ArrayList<>();
+    String dirPath = sc.getRealPath("/party/files");
+
+    for (Part part : files) {
+      if (part.getSize() == 0) {
+        continue;
+      }
+      String filename = UUID.randomUUID().toString();
+      part.write(dirPath + "/" + filename);
+      attachedFiles.add(new AttachedFile(filename));
+    }
+    return attachedFiles;
+  }
+
+  private List<AttachedFile> saveAttachedFiles(MultipartFile[] files)
+      throws IOException, ServletException {
+    List<AttachedFile> attachedFiles = new ArrayList<>();
+    String dirPath = sc.getRealPath("/party/files");
+
+    for (MultipartFile part : files) {
+      if (part.isEmpty()) {
+        continue;
+      }
+
+      System.out.println("filename3 = " + Arrays.toString(files));
+      System.out.println("filename4 = " + files);
+      System.out.println("dirPath = " + dirPath);
+
+      String filename = UUID.randomUUID().toString();
+      part.transferTo(new File(dirPath + "/" + filename));
+      attachedFiles.add(new AttachedFile(filename));
+    }
+    return attachedFiles;
+  }
+
+  // 마이페이지 예약 내역 관리
+  @GetMapping("/myReservationList")
+  public String myReservationList(Model model, HttpSession session) throws Exception {
+    Member loginMember = (Member) session.getAttribute("loginMember");
+    if (loginMember != null) {
+      model.addAttribute("tickets", eventService.getTicketNo(loginMember.getNo()));
+      return "mypage/myReservationList";
+    }
+    return "/auth/login";
   }
 
 }
